==> 2.Data_Cleaning.ipynb <==
import pandas as pd
import numpy as np
import sklearn
from sklearn.model_selection import train_test_split 
from sklearn.preprocessing import LabelEncoder
import seaborn as sns
from sklearn.impute import KNNImputer

df = pd.read_csv(r'data\raw_ckd.csv')
df.head()

print(df.isnull().sum())

It is observed that there are huge number of missing values in all columns. This needs to be handled with an appropriate missing value handling techniques.

df.dtypes

All the string/object dtype columsn needs to be encoded with a suitable encoding. In this case we will be using lable encoding as this is nominal (no order/rank) categorical datatypes

df_cat = df.select_dtypes(include='object')
#col_name = df_cat.columns().to_list()
df_cat 

df_num = df.select_dtypes(exclude='object')
df_num

cat_cols = df_cat.columns.to_list()

for i in cat_cols:
    print ("Unique values in feature:",i,df[i].unique(),'\n')

# Not using direct Label Encoding as data contains NaN values which will be considered as another class
'''
le = LabelEncoder()
df_cat = df_cat.apply(le.fit_transform)
df_cat
'''

cleanup = {"rbc":     {"normal": 1, "abnormal": 0},
           "pc": {"normal": 1, "abnormal": 0},
           "pcc": {"present": 1, "notpresent": 0},
           "ba": {"present": 1, "notpresent": 0},
           "htn": {"yes": 1, "no": 0},
           "dm": {"yes": 1, "no": 0},
           "cad": {"yes": 1, "no": 0},
           "appet": {"good": 1, "poor": 0},
           "pe": {"yes": 1, "no": 0},
           "ane": {"yes": 1, "no": 0},
           "class": {"ckd": 1, "notckd": 0}}

# Encoding Labels with above 'cleanup' notation
df_cat.replace(cleanup,inplace=True)
df_cat

df_final = pd.concat([df_cat,df_num],axis = 1)
df_final

df_final.dtypes

def knnImputerDatasetGenerator(df, neighbors):
    
    imputer_uniform = KNNImputer(n_neighbors=neighbors, weights = 'uniform')
    dataset_uniform = imputer_uniform.fit_transform(df)
    
    return pd.DataFrame(dataset_uniform, columns = df.columns)

# Using KNN imputation for filling NA values. Used different KNN methods with varying no of neighbors considered(n_neighbors)

df_knn_imputed_uniform_3 = knnImputerDatasetGenerator(df_final, 3)
df_knn_imputed_uniform_5 = knnImputerDatasetGenerator(df_final, 5)
df_knn_imputed_uniform_7 = knnImputerDatasetGenerator(df_final, 7)
df_knn_imputed_uniform_9 = knnImputerDatasetGenerator(df_final, 9)    
df_knn_imputed_uniform_11 = knnImputerDatasetGenerator(df_final, 11)
df_mean_imputed = df_final.fillna(df_final.mean())

path = 'data'
df_knn_imputed_uniform_3.to_csv(path + "/df_knn_imputed_uniform_3.csv", index=False)
df_knn_imputed_uniform_5.to_csv(path + "/df_knn_imputed_uniform_5.csv", index=False)
df_knn_imputed_uniform_7.to_csv(path + "/df_knn_imputed_uniform_7.csv", index=False)
df_knn_imputed_uniform_9.to_csv(path + "/df_knn_imputed_uniform_9.csv", index=False)
df_knn_imputed_uniform_11.to_csv(path + "/df_knn_imputed_uniform_11.csv", index=False)
df_mean_imputed.to_csv(path + "/df_mean_imputed.csv", index=False)

### Here we generate 6 different imputed datasets, which will be helful for the next prediction process
