==> 5.After_Feature_Selection.ipynb <==
## Applying Models after Feature Selection

# Importing Libraries
import pandas as pd
import numpy as np
from sklearn.metrics import precision_score, recall_score, f1_score, roc_auc_score, accuracy_score, classification_report
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.impute import KNNImputer

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn.naive_bayes import GaussianNB
from sklearn.neural_network import MLPClassifier

from sklearn.metrics import *
import zipfile
from zipfile import ZipFile
import pickle
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()

# Loading Datasets

df_knn_imputed_uniform_3 = pd.read_csv(r'data\df_knn_imputed_uniform_3_FS.csv')
df_knn_imputed_uniform_5 = pd.read_csv(r'data\df_knn_imputed_uniform_5_FS.csv')
df_knn_imputed_uniform_7 = pd.read_csv(r'data\df_knn_imputed_uniform_7_FS.csv')
df_knn_imputed_uniform_9 = pd.read_csv(r'data\df_knn_imputed_uniform_9_FS.csv')
df_knn_imputed_uniform_11 = pd.read_csv(r'data\df_knn_imputed_uniform_11_FS.csv')
df_mean_imputed = pd.read_csv(r'data\df_mean_imputed_FS.csv')

df_knn_imputed_uniform_3.columns

## Applying all the models(LOG, RF, KNN, SVC, NB and combined model of LOG & RF).

dfs = [df_knn_imputed_uniform_3,df_knn_imputed_uniform_5,
        df_knn_imputed_uniform_7,df_knn_imputed_uniform_9,
        df_knn_imputed_uniform_11,df_mean_imputed]
    
dfs_name = ["knn_imputed_3", "knn_imputed_5",
                "knn_imputed_7","knn_imputed_9",
                "knn_imputed_11","mean_imputed"]

## Logistic Regression model

lr_classifier = LogisticRegression()
def log():
    logistic = {}

    for i, df in enumerate(dfs):
        x = df.drop(['class'], axis=1)
        y = df['class']
        x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.25, shuffle = True, random_state=0)
        
        ss = StandardScaler()
        x_train = ss.fit_transform(x_train)
        x_test = ss.transform(x_test)
        lr_classifier.fit(x_train,y_train)
        
        test_pred = lr_classifier.predict(x_test)
        train_pred = lr_classifier.predict(x_train).round(7)
       
        # print(test_pred)
        train_acc = accuracy_score(y_train, train_pred).round(7)
        test_acc = accuracy_score(y_test, test_pred).round(7)
        
        tn, fp, fn, tp = confusion_matrix(y_test, test_pred).ravel()
#         print(classification_report(y_test, test_pred))
        logistic[dfs_name[i]] = [train_acc, test_acc, tn, fp, fn, tp]
        
    return logistic

pd.DataFrame.from_dict(data = log(),orient='index', columns = ['train_acc', 'test_acc', 'true_neg', 'false_pos', 'false_neg', 'true_pos'])

#### It is observed that for above Logistic Regression FP are between 2 or 3 and FN is zero for KNN imputed datasets. This is good. 
#### But for mean imputed dataset there are FN which is not desirable in CKD prediction use case

## Random Forest Model

rf_classifier = RandomForestClassifier(n_estimators = 500)
def rf():
    rand_forest = {}

    for i, df in enumerate(dfs):
        x = df.drop('class',axis=1)
        y = df['class']
        x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.25, shuffle = True, random_state=42)
        
        ss = StandardScaler()
        x_train = ss.fit_transform(x_train)
        x_test = ss.transform(x_test)
        rf_classifier.fit(x_train,y_train)
        
        test_pred = rf_classifier.predict(x_test)
        train_pred = rf_classifier.predict(x_train)
       
        # print(test_pred)
        train_acc = accuracy_score(y_train, train_pred)
        test_acc = accuracy_score(y_test, test_pred)
        
        tn, fp, fn, tp = confusion_matrix(y_test, test_pred).ravel()
#         print(classification_report(y_test, test_pred))
        rand_forest[dfs_name[i]] = [train_acc, test_acc, tn, fp, fn, tp]
        
    return rand_forest

pd.DataFrame.from_dict(data = rf(),orient='index', columns = ['train_acc', 'test_acc', 'true_neg', 'false_pos', 'false_neg', 'true_pos'])

#### In case of Random Forest, same performance is observed for all datasets. Mean imputed dataset is having perfect score.

## KNN, SVC and Naive Bayes models

def models(x):
    
    if(x==0):
        classifier = KNeighborsClassifier()
    elif(x==1):
        classifier = SVC(kernel='rbf')
    else:
        classifier = GaussianNB()
        
        
    res = {}

    for i, df in enumerate(dfs):
        x = df.drop('class',axis=1)
        y = df['class']
        x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.25, shuffle = True, random_state=0)
        
        ss = StandardScaler()
        x_train = ss.fit_transform(x_train)
        x_test = ss.transform(x_test)
        classifier.fit(x_train,y_train)
        
        test_pred = classifier.predict(x_test)
        train_pred = classifier.predict(x_train)
       
        # print(test_pred)
        train_acc = accuracy_score(y_train, train_pred)
        test_acc = accuracy_score(y_test, test_pred)
        
        tn, fp, fn, tp = confusion_matrix(y_test, test_pred).ravel()
#         print(classification_report(y_test, test_pred))
        res[dfs_name[i]] = [train_acc, test_acc, tn, fp, fn, tp]
        
    return res

for i in range(3):
    result = models(i)
    res = pd.DataFrame.from_dict(data = result, orient='index', columns = ['train_acc', 'test_acc', 'true_neg', 'false_pos', 'false_neg', 'true_pos'])
    if(i==0):
        print("\033[1m",i+1,".PERFORMANCE INDICATORS for KNN \n")
    elif(i==1):
        print(i+1,"\033[1m"+".PERFORMANCE INDICATORS for SVC \n")
    
    else:
       print(i+1,"\033[1m"+".PERFORMANCE INDICATORS for Naive Bayes \n")
#     result = models(i)
    print(res,'\n')

## Combined Model of Logistic Regression & Random Forest

from sklearn.ensemble import VotingClassifier
def combined():

    estimators=[ ('lr', lr_classifier), ('rf', rf_classifier)]

    majority_voting = VotingClassifier(estimators, voting='hard')

    sets = {}

        #dfs: array of imputed df
    dfs = [df_knn_imputed_uniform_3,df_knn_imputed_uniform_5,
            df_knn_imputed_uniform_7,df_knn_imputed_uniform_9,
            df_knn_imputed_uniform_11,df_mean_imputed]

    dfs_name = ["knn_imputed_3", "knn_imputed_5",
                    "knn_imputed_7","knn_imputed_9",
                    "knn_imputed_11","mean_imputed"]

    for i, df in enumerate(dfs):
        x = df.drop('class',axis=1)
        y = df['class']

        x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.25,random_state = 42)

        ss = StandardScaler()
        x_train = ss.fit_transform(x_train)
        x_test = ss.transform(x_test)
        majority_voting.fit(x_train,y_train)

        test_pred = majority_voting.predict(x_test)
        train_pred = majority_voting.predict(x_train)

        # print(test_pred)
        train_acc = accuracy_score(y_train, train_pred)*100
        test_acc = accuracy_score(y_test, test_pred)*100

        tn, fp, fn, tp = confusion_matrix(y_test, test_pred).ravel()
#         print(classification_report(y_test, test_pred))
        sets[dfs_name[i]] = [train_acc, test_acc, tn, fp, fn, tp]
    return sets

pd.DataFrame.from_dict(data = combined(), orient='index', columns = ['train_acc', 'test_acc', 'true_neg', 'false_pos', 'false_neg', 'true_pos'])

## CONCLUSION

Here it is observed that after applying combined model of Logistic regression and Random Forest, we get the accuracy rate more than the individual model of Logistic and Random Forest. So, this combined model is the good enough to apply on CKD prediction
